+++ R/use_cmor_readme.R
@@ -9,7 +9,7 @@
 #'
 #' @export
 use_cmor_readme <- function(data, open = rlang::is_interactive()) {
-  template <- system.file("templates", "cmor-readme", package = "cmor.tools", mustWork = TRUE)
+  template <- system.file("templates", "cmor-readme", package = "CMORprojects", mustWork = TRUE)
 
   template_out <- whisker::whisker.render(readLines(template), data)
   writeLines(template_out, "README.Rmd")